+++ src/main/java/org/ladocuploader/app/cli/MockFtpsClientImpl.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 
 @Slf4j
 @Component
-@Profile("!production")
+@Profile({"dev", "test"})
 public class MockFtpsClientImpl implements FtpsClient {
 
   @Override
